refactor: move LanguageDropdown into separate component
